Fail if attempt is made to use ASYNC directive when scheduler not running.
Previously, the thread's default scheduler would run the synchronous thread
on some other thread than the current thread. This ensured that there
was always a thread available to service SYNC requests even if the
caller had neglected to start the scheduler running.

However, this proved difficult to use in practice because this left the
current thread running asynchronously with respect
to the synchronous thread.

Now, the current thread is always the synchronous thread irrespective
of the source of the scheduler. Instead, we insist that Scheduler.run()
is active before any attempt is made to call the ASYNC directive.

This helps ensure that there is a scheduling loop active that will clear
tasklets that are scheduled with the SYNC directive.